renamed new_ms_inferencing to ms_inferencing
