+++ conf/evaluate/default.yaml
@@ -2,7 +2,7 @@ tasks:
   # - evaluate_model
   # - benchmark_nms
   # - visualize_nms
-  - new_ms_inferencing
+  - ms_inferencing
 
 split: test #val
 save_json: true
