+++ src/evaluate.py
@@ -4,7 +4,7 @@ import logging
 from src.evaluating.evaluate_model import main as evaluate_model_main
 from src.evaluating.benchmark_nms import main as benchmark_nms_main
 from src.evaluating.visualize_nms import main as visualize_nms_main 
-from src.evaluating.new_ms_inferencing import main as new_ms_inferencing_main
+from src.evaluating.ms_inferencing import main as ms_inferencing_main
 
 log = logging.getLogger(__name__)
 
@@ -12,7 +12,7 @@ TASK_REGISTRY = {
     "evaluate_model": evaluate_model_main,
     "benchmark_nms": benchmark_nms_main,
     "visualize_nms": visualize_nms_main,
-    "new_ms_inferencing": new_ms_inferencing_main,
+    "ms_inferencing": ms_inferencing_main,
 }
 
 def main(cfg: DictConfig):
